pyproject.toml
[project]
name = "matt-gpt"
version = "0.1.0"
description = "Personal AI avatar system for authentic digital representation"
authors = [
    {name = "Matt"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "psycopg[binary,pool] (>=3.2.9,<4.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "openai (>=1.100.2,<2.0.0)",
    "dspy-ai (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
